+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Fixed
 ~~~~~
 
 - Not copied ``validate_schema`` parameter in ``BaseSchema.parametrize``. Regression after implementing `#383`_
+- Missing ``app``, ``location`` and ``hooks`` parameters in schema when used with ``BaseSchema.parametrize``. `#416`_
 
 `0.24.2`_ - 2020-02-09
 ----------------------
@@ -707,6 +708,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#416: https://github.com/kiwicom/schemathesis/issues/416
 .. _#412: https://github.com/kiwicom/schemathesis/issues/412
 .. _#410: https://github.com/kiwicom/schemathesis/issues/410
 .. _#407: https://github.com/kiwicom/schemathesis/issues/407